factor out JSONDownload into a separate component
